+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/BSerializer_309901669.cs
@@ -30,7 +30,7 @@ namespace com.wilutions.byps.test.api
 			       /* padding up to multiple of alignment */ + 4;			
 		}
 		
-		public override Object read(Object obj1, BInput bin1, int version)
+		public override Object read(Object obj1, BInput bin1, long version)
 		{
 			BInputBin bin = (BInputBin)bin1;
 			BResult_19 obj = (BResult_19)(obj1 != null ? obj1 : bin.onObjectCreated(new BResult_19()));
