+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/BSerializer_8789515.cs
@@ -18,7 +18,7 @@ namespace com.wilutions.byps.test.api
 		public BSerializer_8789515()
 			: base(8789515) {}		
 		
-		public override Object read(object obj1, BInput bin1, int version) {
+		public override Object read(object obj1, BInput bin1, long version) {
 			BInputBin bin = ((BInputBin)bin1);
 			BBufferBin bbuf = bin.bbuf;
 			Dictionary<float,char> map = new Dictionary<float,char>();
@@ -32,7 +32,7 @@ namespace com.wilutions.byps.test.api
 			return map;
 		}
 		
-		public override void write(Object obj1, BOutput bout1, int version)  {
+		public override void write(Object obj1, BOutput bout1, long version)  {
 			BOutputBin bout = ((BOutputBin)bout1);
 			BBufferBin bbuf = bout.bbuf;
 			Dictionary<float,char> map = (Dictionary<float,char>)obj1;
