+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/ver/BSerializer_313337668.cs
@@ -30,7 +30,7 @@ namespace com.wilutions.byps.test.api.ver
 			       /* padding up to multiple of alignment */ + 4;			
 		}
 		
-		public override void write(Object obj1, BOutput bout1, int version)
+		public override void write(Object obj1, BOutput bout1, long version)
 		{
 			BRequest_EvolveIF_setClient obj = (BRequest_EvolveIF_setClient)obj1;			
 			BOutputBin bout = (BOutputBin)bout1;
@@ -38,7 +38,7 @@ namespace com.wilutions.byps.test.api.ver
 			bout.writeObj(obj._partner, false, null);
 		}
 		
-		public override Object read(Object obj1, BInput bin1, int version)
+		public override Object read(Object obj1, BInput bin1, long version)
 		{
 			BInputBin bin = (BInputBin)bin1;
 			BRequest_EvolveIF_setClient obj = (BRequest_EvolveIF_setClient)(obj1 != null ? obj1 : bin.onObjectCreated(new BRequest_EvolveIF_setClient()));
